INST-75 Add erlang and rabbitmq in the package
Credits to https://github.com/rabbitmq/erlang-rpm for the minimal erlang
distribution required for RabbitMQ. Erlang takes 57MB in the installed
packages (full distribution in ~150MB). RabbitMQ taks 6MB.

The RabbitMQ installation is laid out to mimic the debian package setup.
All binaries are symlinked into /opt/scalr-server/embedded/bin, and the
actual files are in
/opt/scalr-server/embedded/lib/rabbitmq/rabbitmq_server-<version>. This
allows to ship two different versions in a package in the future.