INST-75: RabbitMQ configuration
Add configuration system for rabbitmq

RabbitMQ is disabled by default and must be manually enabled with
rabbitmq[:enable] = true

RabbitMQ is running under a dedicated user, scalr-rabbitmq. This user's
home is /opt/scalr-server/var/lib/rabbitmq, which is the location of the
erlang cookie. As a result, HOME must be set to this directory in order
to run rabbitmqctl.

To run rabbitmqctl, use something like:
 # export PATH=/opt/scalr-server/embedded/bin:$PATH
 # HOME=/opt/scalr-server/var/lib/rabbitmq rabbitmqctl ...

Security

A long random password is generated on the first run for the rabbitmq
user. This password is stored in scalr-server-secrets.json, like the DB
and app passwords.

Network

RabbitMQ binds to two ports, one for management and one for AMQP. Both
use SSL. The agents connect to the AMQP port, so it listens on 0.0.0.0
by default. The management interface binds to 127.0.0.1 by default.
There are several configuration settings related to the network:
 - rabbitmq[:bind_host], rabbitmq[:bind_port]: address to bind the main
   service to
 - rabbitmq[:mgmt_bind_host], rabbitmq[:mgmt_bind_port]: address to bind
   the management interface to.
 - routing[:rabbitmq_host]: the endpoint the agents will connect to.
   Defaults to routing[:endpoint_host]. rabbitmq[:bind_port] is used for
   the port.
 - app[:rabbitmq_host]: the host that the app will connect to (internal
   host). 127.0.0.1 by default. Only one host is currently supported.
   rabbitmq[:bind_port] and rabbitmq[:mgmt_bind_port] are used for the
   ports.

Certificates

The SSL certificates for RabbitMQ are placed in
/opt/scalr-server/etc/rabbitmq/server.{key,crt}. They are generated
automatically by chef during the first reconfigure, with rsa 4096 and
25+ years of validity.